+++ __tests__/rules/substituteRowAtRules.test.js
@@ -15,6 +15,7 @@ it('parses @row', () => {
   const output = `
     article {
       display: flex;
+      flex-wrap: nowrap;
     }
     article > *:nth-child(1) {
       margin-left: 0;
@@ -62,6 +63,103 @@ it('parses @row with specified children count', () => {
   const output = `
     article {
       display: flex;
+      flex-wrap: nowrap;
+    }
+    article > *:nth-child(3n+1) {
+      margin-left: 0;
+    }
+    @media (min-width: 0) {
+      article > * {
+        margin-left: 25px;
+      }
+    }
+    @media (min-width: 740px) {
+      article > * {
+        margin-left: 35px;
+      }
+    }
+    @media (min-width: 1024px) {
+      article > * {
+        margin-left: 50px;
+      }
+    }
+    @media (min-width: 1399px) {
+      article > * {
+        margin-left: 50px;
+      }
+    }
+    @media (min-width: 1900px) {
+      article > * {
+        margin-left: 60px;
+      }
+    }
+  `
+
+  return run(input).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})
+
+it('parses @row with specified children count and wrap specifier', () => {
+  const input = `
+    article {
+      @row 3/wrap;
+    }
+  `
+
+  const output = `
+    article {
+      display: flex;
+      flex-wrap: wrap;
+    }
+    article > *:nth-child(3n+1) {
+      margin-left: 0;
+    }
+    @media (min-width: 0) {
+      article > * {
+        margin-left: 25px;
+      }
+    }
+    @media (min-width: 740px) {
+      article > * {
+        margin-left: 35px;
+      }
+    }
+    @media (min-width: 1024px) {
+      article > * {
+        margin-left: 50px;
+      }
+    }
+    @media (min-width: 1399px) {
+      article > * {
+        margin-left: 50px;
+      }
+    }
+    @media (min-width: 1900px) {
+      article > * {
+        margin-left: 60px;
+      }
+    }
+  `
+
+  return run(input).then(result => {
+    expect(result.css).toMatchCSS(output)
+    expect(result.warnings().length).toBe(0)
+  })
+})
+
+it('parses @row with specified children count and wrap specifier nowrap', () => {
+  const input = `
+    article {
+      @row 3/nowrap;
+    }
+  `
+
+  const output = `
+    article {
+      display: flex;
+      flex-wrap: nowrap;
     }
     article > *:nth-child(3n+1) {
       margin-left: 0;
@@ -117,6 +215,7 @@ it('parses @row under @responsive', () => {
   @media (min-width: 0) and (max-width: 739px) {
     article {
       display: flex;
+      flex-wrap: nowrap;
     }
     article > *:nth-child(3n+1) {
       margin-left: 0;
@@ -150,6 +249,7 @@ it('parses @row as child under @responsive', () => {
   @media (min-width: 0) and (max-width: 739px) {
     .inner article {
       display: flex;
+      flex-wrap: nowrap;
     }
     .inner article > *:nth-child(3n+1) {
       margin-left: 0;
@@ -193,6 +293,7 @@ it('parses @row as child under @responsive w/ advanced bpQuery', () => {
   @media (min-width: 1024px) and (max-width: 1398px), (min-width: 1399px) and (max-width: 1899px), (min-width: 1900px) {
     .inner article {
       display: flex;
+      flex-wrap: nowrap;
     }
     .inner article > *:nth-child(3n+1) {
       margin-left: 0;